Move initialization logic into window

+++ temportal-graphics/src/device/logical/device.rs
@@ -21,8 +21,8 @@ impl Device {
 		}
 	}
 
-	pub fn get_queue(device: Rc<Self>, queue_family_index: u32) -> logical::Queue {
-		let vk = device.get_device_queue(queue_family_index);
+	pub fn get_queue(device: Rc<Self>, queue_family_index: usize) -> logical::Queue {
+		let vk = device.get_device_queue(queue_family_index as u32);
 		logical::Queue::from(device, vk)
 	}
